+++ app/sprinkles/admin/routes/permissions.php
@@ -7,6 +7,8 @@
  * @license   https://github.com/userfrosting/UserFrosting/blob/master/LICENSE.md (MIT License)
  */
 
+use UserFrosting\Sprinkle\Core\Util\NoCache;
+
 /**
  * Routes for administrative permission management.
  */
@@ -15,7 +17,7 @@ $app->group('/permissions', function () {
         ->setName('uri_permissions');
 
     $this->get('/p/{id}', 'UserFrosting\Sprinkle\Admin\Controller\PermissionController:pageInfo');
-})->add('authGuard');
+})->add('authGuard')->add(new NoCache());
 
 $app->group('/api/permissions', function () {
     $this->get('', 'UserFrosting\Sprinkle\Admin\Controller\PermissionController:getList');
@@ -23,4 +25,4 @@ $app->group('/api/permissions', function () {
     $this->get('/p/{id}', 'UserFrosting\Sprinkle\Admin\Controller\PermissionController:getInfo');
 
     $this->get('/p/{id}/users', 'UserFrosting\Sprinkle\Admin\Controller\PermissionController:getUsers');
-})->add('authGuard');
+})->add('authGuard')->add(new NoCache());