+++ app/sprinkles/admin/routes/roles.php
@@ -7,6 +7,8 @@
  * @license   https://github.com/userfrosting/UserFrosting/blob/master/LICENSE.md (MIT License)
  */
 
+use UserFrosting\Sprinkle\Core\Util\NoCache;
+
 /**
  * Routes for administrative role management.
  */
@@ -15,7 +17,7 @@ $app->group('/roles', function () {
         ->setName('uri_roles');
 
     $this->get('/r/{slug}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:pageInfo');
-})->add('authGuard');
+})->add('authGuard')->add(new NoCache());
 
 $app->group('/api/roles', function () {
     $this->delete('/r/{slug}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:delete');
@@ -33,7 +35,7 @@ $app->group('/api/roles', function () {
     $this->put('/r/{slug}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:updateInfo');
 
     $this->put('/r/{slug}/{field}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:updateField');
-})->add('authGuard');
+})->add('authGuard')->add(new NoCache());
 
 $app->group('/modals/roles', function () {
     $this->get('/confirm-delete', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:getModalConfirmDelete');
@@ -43,4 +45,4 @@ $app->group('/modals/roles', function () {
     $this->get('/edit', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:getModalEdit');
 
     $this->get('/permissions', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:getModalEditPermissions');
-})->add('authGuard');
+})->add('authGuard')->add(new NoCache());