+++ app/sprinkles/admin/src/Controller/PermissionController.php
@@ -140,11 +140,9 @@ class PermissionController extends SimpleController
 
         $sprunje = $classMapper->createInstance('permission_user_sprunje', $classMapper, $params);
 
-        $response = $sprunje->toResponse($response);
-
         // Be careful how you consume this data - it has not been escaped and contains untrusted user-supplied content.
         // For example, if you plan to insert it into an HTML DOM, you must escape it on the client side (or use client-side templating).
-        return $response;
+        return $sprunje->toResponse($response);
     }
 
     /**