+++ app/system/Sprinkle/SprinkleManager.php
@@ -12,6 +12,7 @@ namespace UserFrosting\System\Sprinkle;
 use Illuminate\Support\Str;
 use Interop\Container\ContainerInterface;
 use UserFrosting\Support\Exception\FileNotFoundException;
+use UserFrosting\Support\Exception\JsonException;
 
 /**
  * Sprinkle manager class.
@@ -35,12 +36,7 @@ class SprinkleManager
     /**
      * @var string The full absolute base path to the sprinkles directory.
      */
-    protected $sprinklesPath;
-
-    /**
-     * @var string[] Keeps track of a mapping from resource stream names to relative paths.
-     */
-    protected $resourcePaths;
+    protected $sprinklesPath = \UserFrosting\APP_DIR_NAME . \UserFrosting\DS . \UserFrosting\SPRINKLES_DIR_NAME . \UserFrosting\DS;
 
     /**
      * Create a new SprinkleManager object.
@@ -50,7 +46,6 @@ class SprinkleManager
     public function __construct(ContainerInterface $ci)
     {
         $this->ci = $ci;
-        $this->sprinklesPath = \UserFrosting\APP_DIR_NAME . \UserFrosting\DS . \UserFrosting\SPRINKLES_DIR_NAME . \UserFrosting\DS;
     }
 
     /**
@@ -72,7 +67,66 @@ class SprinkleManager
     {
         /** @var \UserFrosting\UniformResourceLocator\ResourceLocator $locator */
         $locator = $this->ci->locator;
-        $locator->registerLocation($sprinkleName, $this->sprinklesPath . $sprinkleName);
+        $locator->registerLocation($sprinkleName, $this->getSprinklePath($sprinkleName));
+    }
+
+    /**
+     * Returns sprinkle base path from name
+     *
+     * @param  string $sprinkleName
+     * @return string
+     */
+    public function getSprinklePath($sprinkleName)
+    {
+        // Get Sprinkle and make sure it exist
+        $sprinkle = $this->getSprinkle($sprinkleName);
+        if (!$sprinkle) {
+            throw new FileNotFoundException("Sprinkle `$sprinkleName` doesn't exist.");
+        }
+
+        // Get path and make sure it exist
+        $path = $this->getSprinklesPath() . $sprinkle;
+        if (!file_exists($path)) {
+            throw new FileNotFoundException("Sprinkle `$sprinkleName` should be found at `$path`, but that directory doesn't exist.");
+        }
+
+        return $path;
+    }
+
+    /**
+     * Returns the sprinkle class
+     *
+     * @param  string $sprinkleName
+     * @return string
+     */
+    protected function getSprinkleClass($sprinkleName)
+    {
+        $className = Str::studly($sprinkleName);
+
+        return $this->getSprinkleClassNamespace($sprinkleName) . "\\$className";
+    }
+
+    /**
+     * Returns the claculated sprinkle namespace
+     * @param  string $sprinkleName
+     * @return string The Sprinkle Namespace
+     */
+    public function getSprinkleClassNamespace($sprinkleName)
+    {
+        $className = Str::studly($sprinkleName);
+
+        return "UserFrosting\\Sprinkle\\$className";
+    }
+
+    /**
+     * Returns the sprinkle service provider class
+     *
+     * @param  string $sprinkleName
+     * @return string
+     */
+    protected function getSprinkleDefaultServiceProvider($sprinkleName)
+    {
+        return $this->getSprinkleClassNamespace($sprinkleName) . '\\ServicesProvider\\ServicesProvider';
     }
 
     /**
@@ -80,13 +134,12 @@ class SprinkleManager
      *
      * Creates an object of a subclass of UserFrosting\System\Sprinkle\Sprinkle if defined for the sprinkle (converting to StudlyCase).
      * Otherwise, returns null.
-     * @param  string $name The name of the Sprinkle to initialize.
+     * @param  string $sprinkleName The name of the Sprinkle to initialize.
      * @return mixed  Sprinkle class instance or null if no such class exist
      */
-    public function bootSprinkle($name)
+    public function bootSprinkle($sprinkleName)
     {
-        $className = Str::studly($name);
-        $fullClassName = "\\UserFrosting\\Sprinkle\\$className\\$className";
+        $fullClassName = $this->getSprinkleClass($sprinkleName);
 
         // Check that class exists.  If not, set to null
         if (class_exists($fullClassName)) {
@@ -124,7 +177,7 @@ class SprinkleManager
      *
      * @param string[] $sprinkleNames
      */
-    public function init($sprinkleNames)
+    public function init(array $sprinkleNames)
     {
         foreach ($sprinkleNames as $sprinkleName) {
             $sprinkle = $this->bootSprinkle($sprinkleName);
@@ -153,12 +206,29 @@ class SprinkleManager
      * Return if a Sprinkle is available
      * Can be used by other Sprinkles to test if their dependencies are met
      *
-     * @param  string $name The name of the Sprinkle
+     * @param  string $sprinkleName The name of the Sprinkle
      * @return bool
      */
-    public function isAvailable($name)
+    public function isAvailable($sprinkleName)
     {
-        return in_array($name, $this->getSprinkleNames());
+        return (bool) $this->getSprinkle($sprinkleName);
+    }
+
+    /**
+     * Find sprinkle value from the sprinkles.json
+     *
+     * @param  string       $sprinkleName
+     * @return string|false Return sprinkle name or false if sprinkle not found
+     */
+    public function getSprinkle($sprinkleName)
+    {
+        $mathches = preg_grep("/^$sprinkleName$/i", $this->getSprinkleNames());
+
+        if (count($mathches) <= 0) {
+            return false;
+        }
+
+        return array_values($mathches)[0];
     }
 
     /**
@@ -174,12 +244,11 @@ class SprinkleManager
     /**
      * Register services for a specified Sprinkle.
      *
-     * @param string $name
+     * @param string $sprinkleName
      */
-    public function registerServices($name)
+    public function registerServices($sprinkleName)
     {
-        $className = Str::studly($name);
-        $fullClassName = "\\UserFrosting\\Sprinkle\\$className\\ServicesProvider\\ServicesProvider";
+        $fullClassName = $this->getSprinkleDefaultServiceProvider($sprinkleName);
 
         // Check that class exists, and register services
         if (class_exists($fullClassName)) {
@@ -187,6 +256,29 @@ class SprinkleManager
             $serviceProvider = new $fullClassName();
             $serviceProvider->register($this->ci);
         }
+    }
+
+    /**
+     * Returns sprinklePath parameter
+     *
+     * @return string
+     */
+    public function getSprinklesPath()
+    {
+        return $this->sprinklesPath;
+    }
+
+    /**
+     * Sets sprinklePath parameter
+     *
+     * @param  string $sprinklesPath
+     * @return static
+     */
+    public function setSprinklesPath($sprinklesPath)
+    {
+        $this->sprinklesPath = $sprinklesPath;
+
+        return $this;
     }
 
     /**
@@ -200,10 +292,16 @@ class SprinkleManager
         $sprinklesFile = @file_get_contents($schemaPath);
 
         if ($sprinklesFile === false) {
-            $errorMessage = "Error: Unable to determine Sprinkle load order.  File '$schemaPath' not found or unable to read. Please create a 'sprinkles.json' file and try again.";
+            $errorMessage = "Error: Unable to determine Sprinkle load order. File '$schemaPath' not found or unable to read. Please create a 'sprinkles.json' file and try again.";
             throw new FileNotFoundException($errorMessage);
         }
 
-        return json_decode($sprinklesFile);
+        // Make sure sprinkle contains valid json
+        if (!$data = json_decode($sprinklesFile)) {
+            $errorMessage = "Error: Unable to determine Sprinkle load order. File '$schemaPath' doesn't contain valid json : " . json_last_error_msg();
+            throw new JsonException($errorMessage);
+        }
+
+        return $data;
     }
 }